requirements.txt
fastapi
uvicorn
boto3
azure-storage-blob
python-dotenv
python-multipart